Rename newfolder to rfid_tracking and fix imports
