+++ deeplabcut/rfid_tracking/__init__.py
@@ -1,0 +1,16 @@
+"""RFID tracking utilities and scripts for DeepLabCut."""
+
+from .convert_detection2tracklets import main as convert_detection2tracklets
+from .match_rfid_to_tracklets import main as match_rfid_to_tracklets
+from .make_video import main as make_video
+from .reconstruct_from_pickle import main as reconstruct_from_pickle
+from . import utils
+
+__all__ = [
+    "convert_detection2tracklets",
+    "match_rfid_to_tracklets",
+    "make_video",
+    "reconstruct_from_pickle",
+    "utils",
+]
+
