+++ deeplabcut/rfid_tracking/reconstruct_from_pickle.py
@@ -19,9 +19,9 @@ import time
 import numpy as np
 import pandas as pd
 
-from utils import (
+from .utils import (
     load_tracklets_pickle, save_pickle_safely, frame_idx_from_key,
-    find_mouse_center_index, body_center_from_arr
+    find_mouse_center_index, body_center_from_arr,
 )
 
 # ================== 配置参数 ==================
