+++ deeplabcut/rfid_tracking/roi_definitions.json
